requirements.txt
discord.py>=2.3.2
yt-dlp>=2024.5.27
python-dotenv
PyNaCl>=1.5.0